Improve typesafety of createVariant using overloads
This separates the types for createVariant into two overloads, one for when a custom property is defined and one when its not.
The result of this is perfect inference of types when used in createRestyleComponent while not requiring the user to define
the generic types if they dont want to.