Stronger type safety for createRestyleFunction (#21)
* Improve type-safety by using generics

* Improve TypeSafety of useRestyle hook

This ensures that the output props from useRestyle are actually typed properly instead of casting them to `any`

* Cleanup / improve typings of useRestyle and RestyleFunction

* Strong type safety for createRestyleFunction

Improve the typings of createRestyleFunction, e.g. the prop value in transform will be strongly typed/inferred.

Also added some other improvements, like removing the need for explicit return value on useRestyle that was actually
throwing away the "style" prop in the returned type. This was done by properly typing filterRestyleProps so everything
could be inferred.

* Fix type error in createRestyleFunction.test.ts